feat: add hooks for app
